SetList (#13)
* SetList

en esta oportunidad cambie las listas contenidas por setter al igual que el idFestival ahora es un entero no un long como en versiones anteriores

* updtade

se realizaron los cambios pedidos

+++ src/dao/FestivalDao.java
@@ -42,6 +42,7 @@ public class FestivalDao {
 		return id;
 	}
 
+	
 	public void actualizar(Festival objeto) {
 		try {
 			iniciaOperacion();
@@ -66,7 +67,7 @@ public class FestivalDao {
 		}
 	}
 
-	public Festival traer(long idFestival) {
+	public Festival traer(int idFestival) {
 		Festival objeto = null;
 		try {
 			iniciaOperacion();
